ref(size_t): switch variables to 32bit

+++ src/core/inc/mem_throt.h
@@ -15,17 +15,17 @@
 typedef struct mem_throt_info {
 	bool is_initialized;
 	bool throttled;			 
-	uint64_t counter_id;
-	uint64_t period_us;
-	uint64_t period_counts;
-	uint64_t budget;
+	size_t counter_id;
+	size_t period_us;
+	size_t period_counts;
+	size_t budget; 
 	int64_t budget_left;
-	uint64_t assign_ratio;
+	size_t assign_ratio;
 }mem_throt_t;
 
 extern size_t global_num_ticket_hypervisor;
 
-void mem_throt_config(uint64_t period_us, uint64_t vm_budget, uint64_t* cpu_ratio);
+void mem_throt_config(size_t period_us, size_t vm_budget, size_t* cpu_ratio);
 
 void mem_throt_init();
